#39439 Move "User consent - allow contact for feedback" to new download page
