+++ GenderPayGap.Core/Enums.cs
@@ -273,7 +273,6 @@ namespace GenderPayGap.Core
         public const string Registrations = "GPG-Registrations.csv";
         public const string RegistrationAddresses = "GPG-RegistrationAddresses.csv";
         public const string UnverifiedRegistrations = "GPG-UnverifiedRegistrations.csv";
-        public const string AllowFeedback = "GPG-UsersToContactForFeedback.csv";
         public const string UnfinishedOrganisations = "GPG-UnfinishedOrgs.csv";
         public const string OrphanOrganisations = "GPG-OrphanOrganisations.csv";
         public const string OrganisationScopes = "GPG-Scopes.csv";
