+++ GenderPayGap.WebJob/Functions/UpdateFiles/Functions.UpdateFile.cs
@@ -33,9 +33,6 @@ namespace GenderPayGap.WebJob
                 case Filenames.UnverifiedRegistrations:
                     await UpdateUnverifiedRegistrationsAsync(log, filePath);
                     break;
-                case Filenames.AllowFeedback:
-                    await UpdateUsersToContactForFeedbackAsync(filePath);
-                    break;
                 case Filenames.OrganisationScopes:
                     await UpdateScopesAsync(filePath);
                     break;
