+++ GenderPayGap.WebUI/Controllers/Admin/AdminDownloadsController.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.IO;
@@ -55,6 +55,35 @@ namespace GenderPayGap.WebUI.Controllers
                 .ToList();
 
             string fileDownloadName = $"Gpg-UserConsent-SendUpdates-{DateTime.Now:yyyy-MM-dd HH:mm}.csv";
+            FileContentResult fileContentResult = CreateCsvDownload(records, fileDownloadName);
+
+            return fileContentResult;
+        }
+
+        [HttpGet("downloads-new/user-consent-allow-contact-for-feedback")]
+        public FileContentResult DownloadUserConsentAllowContactForFeedback()
+        {
+            List<User> users = dataRepository.GetAll<User>()
+                .Where(user => user.Status == UserStatuses.Active)
+                .Where(user => user.UserSettings.Any(us => us.Key == UserSettingKeys.AllowContact && us.Value.ToLower() == "true"))
+                .ToList();
+
+            var records = users.Select(
+                    u => new {
+                        u.Firstname,
+                        u.Lastname,
+                        u.JobTitle,
+                        u.EmailAddress,
+                        u.ContactFirstName,
+                        u.ContactLastName,
+                        u.ContactJobTitle,
+                        u.ContactEmailAddress,
+                        u.ContactPhoneNumber,
+                        u.ContactOrganisation
+                    })
+                .ToList();
+
+            string fileDownloadName = $"Gpg-UserConsent-AllowContactForFeedback-{DateTime.Now:yyyy-MM-dd HH:mm}.csv";
             FileContentResult fileContentResult = CreateCsvDownload(records, fileDownloadName);
 
             return fileContentResult;
