+++ GenderPayGap.WebUI/Views/AdminDownloads/Downloads.cshtml
@@ -49,5 +49,14 @@
             Users who answered "Yes" to "I would like to receive information about webinars, events and new guidance"
         </p>
 
+        <p class="govuk-body">
+            <a href="@Url.Action("DownloadUserConsentAllowContactForFeedback", "AdminDownloads")"
+               class="govuk-link govuk-!-font-weight-bold">
+                Users to contact for feedback
+            </a>
+            <br />
+            Users who answered "Yes" to "I'm happy to be contacted for feedback on this service and take part in gender pay gap surveys"
+        </p>
+
     </div>
 </div>
